pyproject.toml
[tool.poetry]
name = "db-bot"
version = "0.1.0"
description = ""
authors = ["goftok"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.1.0"
sqlalchemy = "^2.0.23"
selenium = "^4.15.2"
python-dotenv = "^1.0.0"
rich = "^13.7.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
